Sync materials with tutorials
