+++ mandelbrot-set-python/06_viewport_pixel.py
@@ -2,7 +2,7 @@ from PIL import Image
 from PIL import ImageEnhance
 
 from mandelbrot_03 import MandelbrotSet
-from viewport_02 import Viewport
+from viewport import Viewport
 
 if __name__ == "__main__":
     print("This might take a while...")
@@ -12,8 +12,8 @@ if __name__ == "__main__":
     image = Image.new(mode="L", size=(512, 512))
     for pixel in Viewport(image, center=-0.7435 + 0.1314j, width=0.002):
         c = complex(pixel)
-        probability = 1 - mandelbrot_set.probability(c, smooth=True)
-        pixel.color = max(0, min(int(probability * 255), 255))
+        stability = 1 - mandelbrot_set.stability(c, smooth=True)
+        pixel.color = int(stability * 255)
 
     enhancer = ImageEnhance.Brightness(image)
     enhancer.enhance(1.25).show()
