+++ mandelbrot-set-python/07_color_palette_matplotlib.py
@@ -1,13 +1,14 @@
+import matplotlib.cm
 from PIL import Image
 
 from mandelbrot_03 import MandelbrotSet
-from viewport_02 import Viewport
+from viewport import Viewport
 
 
 def paint(mandelbrot_set, viewport, palette, smooth):
     for pixel in viewport:
-        probability = mandelbrot_set.probability(complex(pixel), smooth)
-        index = int(min(probability * len(palette), len(palette) - 1))
+        stability = mandelbrot_set.stability(complex(pixel), smooth)
+        index = int(min(stability * len(palette), len(palette) - 1))
         pixel.color = palette[index % len(palette)]
 
 
@@ -20,13 +21,11 @@ def denormalize(palette):
 if __name__ == "__main__":
     print("This might take a while...")
 
-    exterior = [(1, 1, 1)] * 50
-    interior = [(1, 1, 1)] * 5
-    gray_area = [(1 - i / 44,) * 3 for i in range(45)]
-    palette = denormalize(exterior + gray_area + interior)
+    colormap = matplotlib.cm.get_cmap("twilight").colors
+    palette = denormalize(colormap)
 
-    mandelbrot_set = MandelbrotSet(max_iterations=20)
+    mandelbrot_set = MandelbrotSet(max_iterations=512)
     image = Image.new(mode="RGB", size=(512, 512))
-    viewport = Viewport(image, center=-0.75, width=3.5)
+    viewport = Viewport(image, center=-0.7435 + 0.1314j, width=0.002)
     paint(mandelbrot_set, viewport, palette, smooth=True)
     image.show()
