+++ mandelbrot-set-python/10_color_hsb.py
@@ -2,7 +2,7 @@ from PIL import Image
 from PIL.ImageColor import getrgb
 
 from mandelbrot_03 import MandelbrotSet
-from viewport_02 import Viewport
+from viewport import Viewport
 
 
 def hsb(hue_degrees: int, saturation: float, brightness: float):
@@ -19,14 +19,13 @@ if __name__ == "__main__":
     mandelbrot_set = MandelbrotSet(max_iterations=25)
     image = Image.new(mode="RGB", size=(512, 512))
     for pixel in Viewport(image, center=-0.75, width=3.5):
-        p = mandelbrot_set.probability(complex(pixel), smooth=True)
-        probability = max(0, min(p, 1))
+        stability = mandelbrot_set.stability(complex(pixel), smooth=True)
         pixel.color = (
             (0, 0, 0)
-            if probability == 1
+            if stability == 1
             else hsb(
-                hue_degrees=int(probability * 360),
-                saturation=probability,
+                hue_degrees=int(stability * 360),
+                saturation=stability,
                 brightness=1,
             )
         )
