+++ mandelbrot-set-python/requirements.txt
@@ -1,9 +1,13 @@
-cycler==0.10.0
+cycler==0.11.0
+fonttools==4.28.1
 kiwisolver==1.3.2
-matplotlib==3.4.3
-numpy==1.21.2
-Pillow==8.3.2
-pyparsing==2.4.7
+matplotlib==3.5.0
+numpy==1.21.4
+packaging==21.3
+Pillow==8.4.0
+pyparsing==3.0.6
 python-dateutil==2.8.2
-scipy==1.7.1
+scipy==1.7.2
+setuptools-scm==6.3.2
 six==1.16.0
+tomli==1.2.2
