--- mandelbrot-set-python/viewport_02.py
@@ -1,48 +1,0 @@
-# viewport_02.py
-
-from typing import NamedTuple
-
-from PIL import Image
-
-
-class Viewport(NamedTuple):
-    image: Image.Image
-    center: complex
-    width: float
-
-    @property
-    def height(self):
-        return self.scale * self.image.height
-
-    @property
-    def offset(self):
-        return self.center + complex(-self.width, self.height) / 2
-
-    @property
-    def scale(self):
-        return self.width / self.image.width
-
-    def __iter__(self):
-        for y in range(self.image.height):
-            for x in range(self.image.width):
-                yield Pixel(self, x, y)
-
-
-class Pixel(NamedTuple):
-    viewport: Viewport
-    x: int
-    y: int
-
-    @property
-    def color(self):
-        return self.viewport.image.getpixel((self.x, self.y))
-
-    @color.setter
-    def color(self, value):
-        self.viewport.image.putpixel((self.x, self.y), value)
-
-    def __complex__(self):
-        return (
-            complex(self.x, -self.y) * self.viewport.scale
-            + self.viewport.offset
-        )
